KEYS: Generalise system_verify_data() to provide access to internal content
Generalise system_verify_data() to provide access to internal content
through a callback.  This allows all the PKCS#7 stuff to be hidden inside
this function and removed from the PE file parser and the PKCS#7 test key.

If external content is not required, NULL should be passed as data to the
function.  If the callback is not required, that can be set to NULL.

The function is now called verify_pkcs7_signature() to contrast with
verify_pefile_signature() and the definitions of both have been moved into
linux/verification.h along with the key_being_used_for enum.

Change-Id: Ie3de26be692e074aeeb7f332bd033c376e8b5afa
Signed-off-by: David Howells <dhowells@redhat.com>
Signed-off-by: Marco Zanin <mrczn.bb@gmail.com>