block: rename bio bi_rw to bi_opf
Since commit 63a4cc24867d, bio->bi_rw contains flags in the lower
portion and the op code in the higher portions. This means that
old code that relies on manually setting bi_rw is most likely
going to be broken. Instead of letting that brokeness linger,
rename the member, to force old and out-of-tree code to break
at compile time instead of at runtime.

No intended functional changes in this commit.

[B--B: adapted to 4.4 kernel]

Signed-off-by: Jens Axboe <axboe@fb.com>
Signed-off-by: Marco Zanin <mrczn.bb@gmail.com>