PM / devfreq: arm-memlat-mon: Delete event count overflow handling code
The event counts that are u64 are unlikely to overflow before the system is
rebooted. And even if we never reboot the system in 100+ years and the
event counts overflow and we don't handle it, the worst case behavior would
be voting for the wrong memory frequency (bad power/performance) for a few
milliseconds. This extra code is definitely not worth it.

Change-Id: Ia7c8ac3966e96e6bc123bb26dfdba3ccae1e4a4c
Signed-off-by: Saravana Kannan <skannan@codeaurora.org>
Signed-off-by: celtare21 <celtare21@gmail.com>
Signed-off-by: Marco Zanin <mrczn.bb@gmail.com>