net: ipc_router: Fix remote port memory leak
msm_ipc_router_xprt_notify takes a reference to a remote port without
releasing it, causing the memory allocated to the remote port to be
leaked.

Fix the memory leak by putting the reference to the remote port back when
msm_ipc_router_xprt_notify is done with it.

This fixes the following memory leak:
unreferenced object 0xffffffd40b876b00 (size 128):
  comm "mpss_IPCRTR", pid 1196, jiffies 4294880345 (age 79.843s)
  hex dump (first 32 bytes):
    00 01 00 00 00 00 ad de 00 02 00 00 00 00 ad de  ................
    01 00 00 00 01 00 00 00 01 00 00 00 00 00 00 00  ................
  backtrace:
    [<ffffff81d67fbc80>] kmem_cache_alloc_trace+0x218/0x2a0
    [<ffffff81d791f0ac>] ipc_router_create_rport+0x9c/0x1e0
    [<ffffff81d7922dc0>] do_read_data+0x498/0xad8
    [<ffffff81d66c9a40>] kthread_worker_fn+0xa8/0x188
    [<ffffff81d66c9990>] kthread+0xf8/0x100
    [<ffffff81d6682ef0>] ret_from_fork+0x10/0x20
    [<ffffffffffffffff>] 0xffffffffffffffff

Signed-off-by: Sultan Alsawaf <sultan@kerneltoast.com>
Signed-off-by: Marco Zanin <mrczn.bb@gmail.com>